// ccm_web_app_anglr/src/app/core/models/cpt.option.ts
export class CptOption {

    id: number;
    cptOptionId: number;
    name: string;
    code: string;
    toolTip: string;
    description: string;
    sortOrder: number;
    createdBy: number;
    updatedBy: number;
    createdOn: string;
    updatedOn: string;
    isActive: boolean;
    isSelected: boolean;
}